Consolidate parser and formatter
